+++ esp-metadata-generated/src/_generated_esp32h2.rs
@@ -43,10 +43,16 @@ macro_rules! property {
         true
     };
     ("soc.rc_fast_clk_default") => {
-        8000000
+        8500000
     };
     ("soc.rc_fast_clk_default", str) => {
-        stringify!(8000000)
+        stringify!(8500000)
+    };
+    ("soc.rc_slow_clock") => {
+        136000
+    };
+    ("soc.rc_slow_clock", str) => {
+        stringify!(136000)
     };
     ("soc.xtal_frequency") => {
         32