+++ examples/wifi/csi/src/main.rs
@@ -13,7 +13,7 @@ use esp_alloc as _;
 use esp_backtrace as _;
 use esp_hal::{clock::CpuClock, main, rng::Rng, time, timer::timg::TimerGroup};
 use esp_println::println;
-use esp_radio::wifi::{ClientConfiguration, Configuration, CsiConfig, ScanConfig};
+use esp_radio::wifi::{ClientConfig, Config, CsiConfig, ScanConfig};
 use smoltcp::{
     iface::{SocketSet, SocketStorage},
     wire::DhcpOption,
@@ -57,8 +57,8 @@ fn main() -> ! {
     let now = || time::Instant::now().duration_since_epoch().as_millis();
     let stack = Stack::new(iface, device, socket_set, now, rng.random());
 
-    let client_config = Configuration::Client({
-        let mut config = ClientConfiguration::default();
+    let client_config = Config::Client({
+        let mut config = ClientConfig::default();
         config.ssid = SSID.into();
         config.password = PASSWORD.into();
         config