+++ examples/wifi/dhcp/src/main.rs
@@ -24,7 +24,7 @@ use esp_hal::{
     timer::timg::TimerGroup,
 };
 use esp_println::{print, println};
-use esp_radio::wifi::{ClientConfiguration, Configuration, ScanConfig};
+use esp_radio::wifi::{ClientConfig, Config, ScanConfig};
 use smoltcp::{
     iface::{SocketSet, SocketStorage},
     wire::{DhcpOption, IpAddress},
@@ -72,8 +72,8 @@ fn main() -> ! {
         .set_power_saving(esp_radio::config::PowerSaveMode::None)
         .unwrap();
 
-    let client_config = Configuration::Client({
-        let mut config = ClientConfiguration::default();
+    let client_config = Config::Client({
+        let mut config = ClientConfig::default();
         config.ssid = SSID.into();
         config.password = PASSWORD.into();
         config