+++ examples/wifi/static_ip/src/main.rs
@@ -21,7 +21,7 @@ use esp_hal::{
     timer::timg::TimerGroup,
 };
 use esp_println::{print, println};
-use esp_radio::wifi::{ClientConfiguration, Configuration, ScanConfig};
+use esp_radio::wifi::{ClientConfig, Config, ScanConfig};
 use smoltcp::iface::{SocketSet, SocketStorage};
 
 esp_bootloader_esp_idf::esp_app_desc!();
@@ -61,8 +61,8 @@ fn main() -> ! {
     let now = || time::Instant::now().duration_since_epoch().as_millis();
     let mut stack = Stack::new(iface, device, socket_set, now, rng.random());
 
-    let client_config = Configuration::Client({
-        let mut config = ClientConfiguration::default();
+    let client_config = Config::Client({
+        let mut config = ClientConfig::default();
         config.ssid = SSID.into();
         config.password = PASSWORD.into();
         config