+++ xtask/src/main.rs
@@ -12,6 +12,7 @@ use xtask::{
     Package,
     cargo::{CargoAction, CargoArgsBuilder},
     commands::*,
+    update_metadata,
 };
 
 // ----------------------------------------------------------------------------
@@ -171,7 +172,7 @@ fn main() -> Result<()> {
         Cli::LintPackages(args) => lint_packages(&workspace, args),
         Cli::SemverCheck(args) => semver_checks(&workspace, args),
         Cli::CheckChangelog(args) => check_changelog(&workspace, &args.packages, args.normalize),
-        Cli::UpdateMetadata(args) => xtask::update_metadata(&workspace, args.check),
+        Cli::UpdateMetadata(args) => update_metadata(&workspace, args.check),
     }
 }